Combine chance functions and use pytest monkeypatch technique.
